Split service-new into smaller components
