+++ libs/domains/services/feature/src/lib/service-new/service-new.tsx
@@ -1,5 +1,3 @@
-import * as Dialog from '@radix-ui/react-dialog'
-import clsx from 'clsx'
 import posthog from 'posthog-js'
 import { useFeatureFlagEnabled } from 'posthog-js/react'
 import {
@@ -7,740 +5,16 @@ import {
   type CloudProviderEnum,
   type LifecycleTemplateListResponseResultsInner,
 } from 'qovery-typescript-axios'
-import { type ReactElement, cloneElement, useMemo, useState } from 'react'
-import Markdown from 'react-markdown'
-import remarkGfm from 'remark-gfm'
-import { match } from 'ts-pattern'
-import { type ServiceType } from '@qovery/domains/services/data-access'
-import { Badge, Button, ExternalLink, Heading, Icon, InputSearch, Link, Section } from '@qovery/shared/ui'
+import { useMemo, useState } from 'react'
+import { ExternalLink, Heading, Icon, InputSearch, Section } from '@qovery/shared/ui'
 import { useSupportChat } from '@qovery/shared/util-hooks'
-import { twMerge } from '@qovery/shared/util-js'
-import { TemplateIds } from '@qovery/shared/util-services'
-import { useBlueprintCatalogServiceReadme } from '../hooks/use-blueprint-catalog-service-readme/use-blueprint-catalog-service-readme'
-import { ServiceIcons } from '../service-icon/service-icon'
-import {
-  type ServiceTemplateOptionType,
-  type ServiceTemplateType,
-  type TagsEnum,
-  serviceTemplates,
-} from './service-templates'
+import { BlueprintCard } from './blueprint-card/blueprint-card'
+import { BlueprintDetailsPanel } from './blueprint-details-panel/blueprint-details-panel'
+import { Card, CardService, SectionByTag, type ServiceBlock } from './service-card/service-card'
+import { buildCreateFlowPathForType, getCreateFlowPath, getServicesPath } from './service-new-utils/service-new-utils'
+import { serviceTemplates } from './service-templates'
 
 const CloudFormationIcon = '/assets/devicon/cloudformation.svg'
-
-const getEnvironmentBasePath = (organizationId: string, projectId: string, environmentId: string) =>
-  `/organization/${organizationId}/project/${projectId}/environment/${environmentId}`
-
-const getServicesPath = (organizationId: string, projectId: string, environmentId: string, subPath: string) =>
-  `${getEnvironmentBasePath(organizationId, projectId, environmentId)}${subPath}`
-
-const CREATE_FLOW_SLUG_BY_TYPE: Partial<Record<ServiceType, string>> = {
-  APPLICATION: 'application',
-  CONTAINER: 'container',
-  DATABASE: 'database',
-  HELM: 'helm',
-  JOB: 'lifecycle-job',
-  LIFECYCLE_JOB: 'lifecycle-job',
-  CRON_JOB: 'cron-job',
-  TERRAFORM: 'terraform',
-}
-
-const CREATE_FLOW_SLUGS = new Set(Object.values(CREATE_FLOW_SLUG_BY_TYPE))
-
-/**
- * Path = flow slug (from config), template = preset id in query.
- * Works for card-by-type (Application > Next.js) and card-by-template (Apache Kafka > Container).
- * */
-function getCreateFlowPathParams(
-  type: ServiceType,
-  parentSlug: string,
-  slug: string
-): { flowSlug: string; templateSlug: string | undefined } {
-  if (slug === 'current') {
-    const flowSlug = CREATE_FLOW_SLUGS.has(parentSlug) ? parentSlug : CREATE_FLOW_SLUG_BY_TYPE[type] ?? parentSlug
-    const templateSlug = CREATE_FLOW_SLUGS.has(parentSlug) ? undefined : parentSlug
-    return { flowSlug, templateSlug }
-  }
-  const flowSlug = CREATE_FLOW_SLUGS.has(parentSlug) ? parentSlug : slug
-  const templateSlug = CREATE_FLOW_SLUGS.has(parentSlug) ? (CREATE_FLOW_SLUGS.has(slug) ? undefined : slug) : parentSlug
-  return { flowSlug, templateSlug }
-}
-
-/** Builds /service/create/:flowSlug with optional ?template= for presets. Use for any type listed in CREATE_FLOW_SLUG_BY_TYPE. */
-function getCreateFlowPath(flowSlug: string, templateSlug?: string): string {
-  const path = `/service/create/${flowSlug}`
-  if (templateSlug && templateSlug !== 'current') {
-    return `${path}?template=${encodeURIComponent(templateSlug)}`
-  }
-  return path
-}
-
-function buildCreateFlowPathForType(type: ServiceType, parentSlug: string, slug: string): string | undefined {
-  const flowSlug = CREATE_FLOW_SLUG_BY_TYPE[type]
-  if (!flowSlug) return undefined
-
-  if (type === 'DATABASE' || type === 'LIFECYCLE_JOB') {
-    const templateSlug = parentSlug === flowSlug || parentSlug === 'current' ? undefined : parentSlug
-    const optionSlug = slug === 'current' ? undefined : slug
-    const path = getCreateFlowPath(flowSlug, templateSlug)
-
-    if (optionSlug) {
-      return `${path}${templateSlug ? '&' : '?'}option=${encodeURIComponent(optionSlug)}`
-    }
-
-    return path
-  }
-
-  const { flowSlug: resolvedFlowSlug, templateSlug } = getCreateFlowPathParams(type, parentSlug, slug)
-  return getCreateFlowPath(resolvedFlowSlug, templateSlug)
-}
-
-function Card({
-  title,
-  description,
-  icon,
-  link,
-  onClick,
-  disabledCTA,
-  badge,
-}: {
-  title: string
-  description: string
-  icon: ReactElement
-  link?: string
-  onClick?: () => void
-  disabledCTA?: ReactElement
-  badge?: string
-}) {
-  const Wrapper = ({ children }: { children: ReactElement }) => {
-    const className = twMerge(
-      'flex cursor-pointer items-center justify-between gap-5 rounded border border-neutral px-5 py-4 transition [box-shadow:0px_2px_8px_-1px_rgba(27,36,44,0.08),0px_2px_2px_-1px_rgba(27,36,44,0.04)]',
-      disabledCTA ? 'border-neutral bg-surface-neutral-subtle' : 'hover:bg-surface-neutral-subtle'
-    )
-
-    if (onClick) {
-      return (
-        <div onClick={onClick} className={className}>
-          {children}
-        </div>
-      )
-    }
-
-    if (!link) {
-      return <div className={className}>{children}</div>
-    }
-
-    return (
-      // @ts-expect-error-next-line TODO new-nav : Route strings need to be updated using the next typed routes
-      <Link to={link} className={className}>
-        {children}
-      </Link>
-    )
-  }
-
-  return (
-    <Wrapper>
-      <>
-        <div className="space-y-2">
-          <div className="space-y-1">
-            <div className="flex items-center gap-2">
-              <h3 className="text-ssm font-medium text-neutral">{title}</h3>
-              {badge && (
-                <Badge
-                  radius="full"
-                  variant="surface"
-                  color="purple"
-                  size="sm"
-                  className="h-4 border-transparent bg-surface-accent1-component px-1 text-[8px]  text-accent1"
-                >
-                  {badge}
-                </Badge>
-              )}
-            </div>
-            <p className="max-w-96 text-xs text-neutral-subtle">{description}</p>
-          </div>
-          {disabledCTA}
-        </div>
-        {icon}
-      </>
-    </Wrapper>
-  )
-}
-
-/** Types listed in CREATE_FLOW_SLUG_BY_TYPE use /service/create/:flowSlug?template=; others use legacy URLs until migrated. */
-const servicePathSuffix = (type: ServiceType, parentSlug: string, slug: string) =>
-  match(type)
-    .with('APPLICATION', 'CONTAINER', 'DATABASE', 'HELM', 'TERRAFORM', 'LIFECYCLE_JOB', () =>
-      buildCreateFlowPathForType(type, parentSlug, slug)
-    )
-    .with('JOB', 'CRON_JOB', () => undefined)
-    .otherwise(() => buildCreateFlowPathForType(type, parentSlug, slug))
-
-interface CardOptionProps extends ServiceTemplateOptionType {
-  parentSlug: string
-  organizationId: string
-  projectId: string
-  environmentId: string
-  isTerraformFeatureFlag?: boolean
-  onUpgradePlanClick?: () => void
-}
-
-function CardOption({
-  parentSlug,
-  slug,
-  icon,
-  title,
-  description,
-  type,
-  recommended,
-  badge,
-  template_id,
-  icon_uri,
-  organizationId,
-  projectId,
-  environmentId,
-  isTerraformFeatureFlag,
-  onUpgradePlanClick,
-}: CardOptionProps) {
-  const iconClassName = ServiceIcons[icon_uri].className
-  const pathSuffix = servicePathSuffix(type, parentSlug, slug)
-  const to = pathSuffix ? getServicesPath(organizationId, projectId, environmentId, pathSuffix) : undefined
-
-  const isTerraformUpgrade = template_id === TemplateIds.TERRAFORM && !isTerraformFeatureFlag && onUpgradePlanClick
-
-  if (isTerraformUpgrade) {
-    return (
-      <div
-        role="button"
-        tabIndex={0}
-        onClick={onUpgradePlanClick}
-        onKeyDown={(e) => e.key === 'Enter' && onUpgradePlanClick()}
-        className="flex cursor-pointer items-start gap-3 rounded-sm border border-neutral bg-surface-neutral-subtle p-3 transition hover:bg-surface-neutral-componentHover"
-      >
-        <img className={twMerge('mt-1 select-none', iconClassName)} width={24} height={24} src={icon} alt={title} />
-        <span className="flex flex-1 flex-col gap-1">
-          <span className="inline-flex items-center gap-2 text-ssm font-medium text-neutral">
-            {title}
-            {badge && (
-              <Badge
-                radius="full"
-                variant="surface"
-                color="purple"
-                size="sm"
-                className="h-4 border-transparent bg-surface-accent1-component px-1 text-[8px]  text-accent1"
-              >
-                {badge}
-              </Badge>
-            )}
-          </span>
-          <span className="inline-block text-xs text-neutral-subtle">{description}</span>
-          <p className="cursor-pointer text-xs font-medium text-neutral-subtle">
-            Upgrade your plan <Icon iconName="chevron-right" className="ml-1 text-2xs" />
-          </p>
-        </span>
-      </div>
-    )
-  }
-
-  if (!to) return null
-
-  return (
-    <Link
-      // @ts-expect-error-next-line TODO new-nav : Route strings need to be updated using the next typed routes
-      to={to}
-      className="flex items-start gap-3 rounded-sm border border-neutral bg-surface-neutral-component p-3 transition hover:bg-surface-neutral-componentHover"
-      onClick={() =>
-        posthog.capture('select-service', {
-          qoveryServiceType: type,
-          selectedServiceType: parentSlug,
-          selectedServiceSubType: slug,
-        })
-      }
-    >
-      <img className={twMerge('mt-1 select-none', iconClassName)} width={24} height={24} src={icon} alt={title} />
-      <span>
-        <span className="inline-flex items-center gap-2 text-ssm font-medium text-neutral">
-          {title}
-          {badge && (
-            <Badge
-              radius="full"
-              variant="surface"
-              color="purple"
-              size="sm"
-              className="h-4 border-transparent bg-surface-accent1-component px-1 text-[8px]  text-accent1"
-            >
-              {badge}
-            </Badge>
-          )}
-          {recommended && (
-            <span className="relative -top-0.5 inline-block rounded bg-surface-brand-solid px-1 text-2xs text-neutralInvert">
-              Fastest
-            </span>
-          )}
-        </span>
-        <span className="inline-block text-xs text-neutral-subtle">{description}</span>
-      </span>
-      <span className="flex h-full items-center pr-1">
-        <Icon iconName="chevron-right" className="text-xs text-neutral-subtle" />
-      </span>
-    </Link>
-  )
-}
-
-function CardService({
-  title,
-  icon,
-  description,
-  slug,
-  options,
-  type,
-  link,
-  availableTemplates,
-  organizationId,
-  projectId,
-  environmentId,
-  cloudProvider,
-  icon_uri,
-  isTerraformFeatureFlag,
-  onUpgradePlanClick,
-}: Omit<ServiceTemplateType, 'cloud_provider'> & {
-  cloud_provider?: CloudProviderEnum | string
-  availableTemplates: LifecycleTemplateListResponseResultsInner[]
-  organizationId: string
-  projectId: string
-  environmentId: string
-  cloudProvider?: CloudProviderEnum | string
-  isTerraformFeatureFlag: boolean
-  onUpgradePlanClick?: () => void
-}) {
-  const [expanded, setExpanded] = useState(false)
-  const iconClassName = icon_uri ? ServiceIcons[icon_uri].className : undefined
-
-  if (options && !slug) {
-    return null
-  }
-
-  if (options) {
-    return (
-      <div
-        onClick={() => setExpanded(true)}
-        className={clsx({
-          'flex cursor-pointer items-center gap-6 rounded border border-neutral p-5 transition [box-shadow:0px_2px_8px_-1px_rgba(27,36,44,0.08),0px_2px_2px_-1px_rgba(27,36,44,0.04)] hover:bg-surface-neutral-subtle':
-            true,
-          'col-span-3 bg-surface-neutral-subtle p-6': expanded,
-        })}
-      >
-        {expanded ? (
-          <div className="flex w-full flex-col gap-8">
-            <div className="relative flex gap-6">
-              <img
-                className={twMerge('select-none', iconClassName)}
-                width={52}
-                height={52}
-                src={icon as string}
-                alt={title}
-              />
-              <div>
-                <h3 className="mb-1 text-base font-medium text-neutral">{title}</h3>
-                <p className="max-w-96 text-ssm text-neutral-subtle">{description}</p>
-              </div>
-              <Button
-                className="absolute right-0 top-0"
-                color="neutral"
-                variant="surface"
-                iconOnly
-                onClick={(e) => {
-                  e.stopPropagation()
-                  setExpanded(false)
-                }}
-              >
-                <Icon iconName="xmark" />
-              </Button>
-            </div>
-            <div className="grid grid-cols-3 gap-2">
-              {options
-                .filter((c) => {
-                  if (c.template_id === TemplateIds.TERRAFORM && !isTerraformFeatureFlag) return true
-                  if (c.template_id) {
-                    return availableTemplates.reduce((acc, template) => c.template_id === template.id || acc, false)
-                  }
-                  return true
-                })
-                .filter((c) => c.cloud_provider === cloudProvider || !c.cloud_provider)
-                .sort((a, b) => {
-                  if (a.recommended && !b.recommended) return -1
-                  if (!a.recommended && b.recommended) return 1
-                  return a.title.localeCompare(b.title)
-                })
-                .map((props) => (
-                  <CardOption
-                    key={props.slug}
-                    parentSlug={slug ?? ''}
-                    organizationId={organizationId}
-                    projectId={projectId}
-                    environmentId={environmentId}
-                    isTerraformFeatureFlag={isTerraformFeatureFlag}
-                    onUpgradePlanClick={onUpgradePlanClick}
-                    {...props}
-                  />
-                ))}
-            </div>
-          </div>
-        ) : (
-          <>
-            <div className="flex h-full w-60 flex-col justify-between gap-2">
-              <div>
-                <h3 className="mb-1 text-ssm font-medium text-neutral">{title}</h3>
-                <p className="text-xs text-neutral-subtle">{description}</p>
-              </div>
-              <p className="text-xs font-medium text-neutral-subtle">
-                Click to select an option <Icon iconName="chevron-right" className="ml-1 text-2xs" />
-              </p>
-            </div>
-            <span className="relative">
-              {typeof icon === 'string' ? (
-                <img className={twMerge('max-h-10 w-14 select-none', iconClassName)} src={icon} alt={title} />
-              ) : (
-                cloneElement(icon as ReactElement, {
-                  className: twMerge('w-10', iconClassName),
-                })
-              )}
-            </span>
-          </>
-        )}
-      </div>
-    )
-  }
-
-  const pathSuffix = link ?? (type && slug ? servicePathSuffix(type, slug, 'current') : undefined)
-  const to = pathSuffix ? getServicesPath(organizationId, projectId, environmentId, pathSuffix) : undefined
-
-  if (!to) return null
-
-  return (
-    <Link
-      // @ts-expect-error-next-line TODO new-nav : Route strings need to be updated using the next typed routes
-      to={to}
-      className="flex gap-6 rounded border border-neutral p-5 transition [box-shadow:0px_2px_8px_-1px_rgba(27,36,44,0.08),0px_2px_2px_-1px_rgba(27,36,44,0.04)] hover:bg-surface-neutral-subtle"
-      onClick={() =>
-        posthog.capture('select-service', {
-          qoveryServiceType: type,
-          selectedServiceSubType: slug,
-        })
-      }
-    >
-      <div className="w-60">
-        <h3 className="mb-1 text-ssm font-medium text-neutral">{title}</h3>
-        <p className="text-xs text-neutral-subtle">{description}</p>
-      </div>
-      <div className="flex items-center">
-        {typeof icon === 'string' ? (
-          <img className={twMerge('max-h-10 w-14 select-none', iconClassName)} src={icon} alt={title} />
-        ) : (
-          cloneElement(icon as ReactElement, {
-            className: twMerge('w-10', iconClassName),
-          })
-        )}
-      </div>
-    </Link>
-  )
-}
-
-function SectionByTag({
-  title,
-  description,
-  tag,
-  cloudProvider,
-  availableTemplates,
-  organizationId,
-  projectId,
-  environmentId,
-  isTerraformFeatureFlag,
-  onUpgradePlanClick,
-}: {
-  title: string
-  tag: keyof typeof TagsEnum
-  cloudProvider?: CloudProviderEnum | string
-  description?: string
-  availableTemplates: LifecycleTemplateListResponseResultsInner[]
-  organizationId: string
-  projectId: string
-  environmentId: string
-  isTerraformFeatureFlag: boolean
-  onUpgradePlanClick?: () => void
-}) {
-  return (
-    <Section>
-      <Heading className="mb-1">{title}</Heading>
-      {description && <p className="text-xs text-neutral-subtle">{description}</p>}
-      <div className="mt-5 grid grid-cols-3 gap-4">
-        {serviceTemplates
-          .filter((c) => c.cloud_provider === cloudProvider || !c.cloud_provider)
-          .filter(({ tag: t }) => t === tag)
-          .sort((a, b) => a.title.localeCompare(b.title))
-          .map((service) => (
-            <CardService
-              key={service.title}
-              availableTemplates={availableTemplates}
-              organizationId={organizationId}
-              projectId={projectId}
-              environmentId={environmentId}
-              cloudProvider={cloudProvider}
-              isTerraformFeatureFlag={isTerraformFeatureFlag}
-              onUpgradePlanClick={onUpgradePlanClick}
-              {...service}
-            />
-          ))}
-      </div>
-    </Section>
-  )
-}
-
-type ServiceBlock = {
-  title: string
-  description: string
-  icon: ReactElement
-  cloud_provider?: CloudProviderEnum | string
-  link?: string
-  onClick?: () => void
-  disabledCTA?: ReactElement
-  badge?: string
-}
-
-function BlueprintCard({
-  blueprint,
-  onViewDetails,
-}: {
-  blueprint: BlueprintItem
-  onViewDetails: (blueprint: BlueprintItem) => void
-}) {
-  return (
-    <section className="flex h-full flex-col gap-4 rounded-lg border border-neutral bg-surface-neutral p-4 [box-shadow:0px_0px_4px_0px_rgba(0,0,0,0.01),0px_2px_3px_0px_rgba(0,0,0,0.02)]">
-      <div className="flex flex-1 flex-col gap-3">
-        <img className="h-9 w-9 rounded" src={blueprint.icon} alt="" aria-hidden="true" />
-        <div className="flex flex-col gap-1">
-          <h3 className="text-sm font-medium leading-5 text-neutral">{blueprint.name}</h3>
-          <p className="text-sm leading-5 text-neutral-subtle">{blueprint.description}</p>
-        </div>
-      </div>
-      <div className="mt-auto flex items-center gap-1">
-        <Button type="button" variant="outline" color="neutral" size="sm">
-          Deploy
-        </Button>
-        <Button type="button" variant="plain" color="neutral" size="sm" onClick={() => onViewDetails(blueprint)}>
-          View details
-        </Button>
-      </div>
-    </section>
-  )
-}
-
-function formatBlueprintProvider(provider: string) {
-  const normalizedProvider = provider.toLowerCase()
-
-  return match(normalizedProvider)
-    .with('aws', () => 'AWS')
-    .with('scw', () => 'Scaleway')
-    .with('gcp', () => 'GCP')
-    .with('azure', () => 'Azure')
-    .otherwise(() => provider.toUpperCase())
-}
-
-function getBlueprintRepositoryName({ provider, serviceFamily }: BlueprintItem) {
-  return `qovery-blueprints/${serviceFamily || provider}`
-}
-
-function getBlueprintRepositoryUrlFromReadme(readme: string | undefined, repositoryName: string) {
-  if (!readme) return undefined
-
-  const [organizationName, repositorySlug] = repositoryName.split('/')
-  const repositoryUrlPattern = new RegExp(
-    `https://github\\.com/${organizationName}/${repositorySlug}(?:[\\w\\-/?.#=&%]*)?`,
-    'i'
-  )
-
-  return readme.match(repositoryUrlPattern)?.[0]
-}
-
-function BlueprintReadmeContent({ children }: { children: string }) {
-  return (
-    <Markdown
-      remarkPlugins={[remarkGfm]}
-      components={{
-        h1: ({ node, ...props }) => <h1 className="mb-4 text-2xl font-medium leading-8 text-neutral" {...props} />,
-        h2: ({ node, ...props }) => <h2 className="mb-3 mt-6 text-lg font-medium leading-7 text-neutral" {...props} />,
-        h3: ({ node, ...props }) => (
-          <h3 className="mb-2 mt-5 text-base font-medium leading-6 text-neutral" {...props} />
-        ),
-        p: ({ node, ...props }) => <p className="my-3 text-sm leading-6 text-neutral" {...props} />,
-        a: ({ node, ...props }) => (
-          <a className="font-medium text-brand hover:underline" target="_blank" rel="noopener noreferrer" {...props} />
-        ),
-        ul: ({ node, ...props }) => <ul className="my-3 list-disc space-y-1 pl-5 text-sm text-neutral" {...props} />,
-        ol: ({ node, ...props }) => <ol className="my-3 list-decimal space-y-1 pl-5 text-sm text-neutral" {...props} />,
-        code: ({ node, className, ...props }) => (
-          <code
-            className={twMerge('rounded bg-surface-neutral-subtle px-1 py-0.5 font-mono text-xs', className)}
-            {...props}
-          />
-        ),
-        pre: ({ node, ...props }) => (
-          <pre
-            className="my-4 overflow-auto rounded border border-neutral bg-surface-neutral-subtle p-4 text-xs text-neutral"
-            {...props}
-          />
-        ),
-        table: ({ node, ...props }) => (
-          <div className="my-4 overflow-x-auto rounded border border-neutral">
-            <table className="w-full border-collapse text-sm text-neutral" {...props} />
-          </div>
-        ),
-        th: ({ node, ...props }) => (
-          <th className="border-b border-r border-neutral px-3 py-2 text-left font-medium" {...props} />
-        ),
-        td: ({ node, ...props }) => <td className="border-b border-r border-neutral px-3 py-2" {...props} />,
-      }}
-    >
-      {children}
-    </Markdown>
-  )
-}
-
-function BlueprintReadmeState({
-  readme,
-  isLoading,
-  isError,
-}: {
-  readme?: string
-  isLoading: boolean
-  isError: boolean
-}) {
-  if (isLoading) {
-    return (
-      <div className="flex min-h-[240px] items-center justify-center gap-2 text-sm text-neutral-subtle">
-        <Icon iconName="loader" className="animate-spin" />
-        Loading blueprint details...
-      </div>
-    )
-  }
-
-  if (isError) {
-    return (
-      <div className="flex min-h-[240px] items-center justify-center text-sm text-neutral-subtle">
-        Unable to load blueprint details.
-      </div>
-    )
-  }
-
-  if (!readme) {
-    return (
-      <div className="flex min-h-[240px] items-center justify-center text-sm text-neutral-subtle">
-        No blueprint details available.
-      </div>
-    )
-  }
-
-  return <BlueprintReadmeContent>{readme}</BlueprintReadmeContent>
-}
-
-function BlueprintDetailsPanel({
-  blueprint,
-  organizationId,
-  open,
-  onOpenChange,
-}: {
-  blueprint: BlueprintItem | null
-  organizationId: string
-  open: boolean
-  onOpenChange: (open: boolean) => void
-}) {
-  const serviceVersion = blueprint?.majorVersions[0]?.serviceVersion ?? ''
-  const {
-    data: readme,
-    isLoading,
-    isError,
-  } = useBlueprintCatalogServiceReadme({
-    organizationId,
-    provider: blueprint?.provider ?? '',
-    serviceFamily: blueprint?.serviceFamily ?? '',
-    serviceVersion,
-    enabled: open && Boolean(blueprint),
-  })
-
-  if (!blueprint) return null
-
-  const provider = formatBlueprintProvider(blueprint.provider)
-  const repositoryName = getBlueprintRepositoryName(blueprint)
-  const repositoryUrl = getBlueprintRepositoryUrlFromReadme(readme, repositoryName)
-  const repositoryBadge = (
-    <Badge size="sm" color="neutral" variant="outline" className="gap-1">
-      <Icon name="GITHUB" width={12} height={12} />
-      {repositoryName}
-    </Badge>
-  )
-
-  return (
-    <Dialog.Root open={open} onOpenChange={onOpenChange}>
-      <Dialog.Portal>
-        <Dialog.Overlay className="fixed inset-0 z-overlay bg-background-overlay" />
-        <Dialog.Content className="fixed bottom-0 right-0 top-0 z-modal flex w-[940px] max-w-[calc(100vw-32px)] flex-col overflow-hidden border-l border-neutral bg-background focus:outline-none">
-          <div className="flex-1 overflow-auto px-6 pb-24 pt-6">
-            <div className="mb-8 flex flex-col gap-5">
-              <div className="flex flex-col gap-2">
-                <Dialog.Title className="flex items-center gap-3 pr-8 text-2xl font-medium leading-8 text-neutral">
-                  <img className="h-8 w-8 rounded" src={blueprint.icon} alt="" aria-hidden="true" />
-                  <span>{blueprint.name}</span>
-                </Dialog.Title>
-                <Dialog.Description className="text-sm leading-5 text-neutral-subtle">
-                  {blueprint.description}
-                </Dialog.Description>
-              </div>
-              <div className="flex flex-wrap items-center gap-2">
-                {repositoryUrl ? (
-                  <ExternalLink href={repositoryUrl} withIcon={false} aria-label={`Open ${repositoryName} repository`}>
-                    {repositoryBadge}
-                  </ExternalLink>
-                ) : (
-                  repositoryBadge
-                )}
-                <Badge size="sm" color="neutral" variant="outline" className="gap-1">
-                  {provider}
-                </Badge>
-                {serviceVersion && (
-                  <Badge size="sm" color="neutral" variant="outline" className="gap-1">
-                    v{serviceVersion}
-                  </Badge>
-                )}
-              </div>
-            </div>
-
-            <div className="rounded border border-neutral bg-surface-neutral p-5">
-              <BlueprintReadmeState readme={readme} isLoading={isLoading} isError={isError} />
-            </div>
-          </div>
-
-          <Dialog.Close asChild>
-            <button
-              type="button"
-              className="absolute right-2 top-2 flex h-6 w-6 items-center justify-center rounded text-neutral-subtle hover:bg-surface-neutral-subtle hover:text-neutral"
-              aria-label="Close blueprint details"
-            >
-              <Icon iconName="xmark" />
-            </button>
-          </Dialog.Close>
-
-          <div className="absolute bottom-0 left-0 right-0 flex items-center justify-end gap-2 border-t border-neutral bg-background px-6 py-4">
-            <Button type="button" variant="plain" color="neutral" size="lg" onClick={() => onOpenChange(false)}>
-              Cancel
-            </Button>
-            <Button type="button" color="brand" size="lg">
-              Deploy blueprint
-            </Button>
-          </div>
-        </Dialog.Content>
-      </Dialog.Portal>
-    </Dialog.Root>
-  )
-}
 
 export interface ServiceNewProps {
   organizationId: string
