+++ libs/shared/util-js/src/index.ts
@@ -7,6 +7,7 @@ export * from './lib/custom-tw-merge'
 export * from './lib/environment-variable-file'
 export * from './lib/format-metric'
 export * from './lib/format-cron-expression'
+export * from './lib/format-cloud-provider'
 export * from './lib/merge-deployment-services'
 export * from './lib/object-flattener'
 export * from './lib/parse-dotenv'
